New parameters for assimilator.  Details:
Added installTrove and projectLabel as parameters to assimilator job so we can
install software in the same job/event without having to round trip to the
server and rely on the "migrateSoftware" eventing, which is a couple of hops
away and as a result more error-prone and fun to debug.   This installs
the software sooner.  We may also want to set the projectLabel into the
search path (to be added).